stronger typing (cell), better encaps, but indexing still broken

+++ IP/cpp/iProlog/Integer.h
@@ -8,10 +8,10 @@ namespace iProlog {
 	using namespace std;
 	class Integer {
 	public:
-		size_t i;
-		inline Integer(size_t x) { i = x;  }
+		int i;
+		inline Integer(int x) { i = x;  }
 		inline Integer() { i = 0; }
-		inline Integer operator =(size_t x) { i = x;  return i; }
+		inline Integer operator =(int x) { i = x;  return i; }
 		bool operator ==(const Integer x) const { return i == x.i;  }
 		bool equal_to(const Integer x) const { return i == x.i;  }
 		inline operator unsigned long () { return (unsigned long)i; }
